handle streamlit refresh token roation

+++ ui/api/api.py
@@ -61,10 +61,31 @@ def get_token(username, password):
     }
     response = requests.post(url, data=form_data)
     if response.status_code == status.HTTP_200_OK:
+        st.session_state.refresh_token = response.json().get("jwt_refresh_token")
+        st.session_state.access_token_expires_at = time.time() + response.json().get("expires_in")
         return response.json().get("jwt_token")
     else:
         st.toast(response.json().get("detail"), icon=":material/error:")
         return None
+
+
+def refresh_tokens():
+    url = f"{BASE_URL}/auth/refresh"
+    payload = {"refresh_token": st.session_state.refresh_token}
+    response = requests.post(url, json=payload)
+    if response.status_code == status.HTTP_200_OK:
+        data = response.json()
+
+        st.session_state.access_token = data.get("jwt_token")
+        st.session_state.refresh_token = data.get("jwt_refresh_token")
+        st.session_state.access_token_expires_at = time.time() + data.get("expires_in")
+    else:
+        st.toast(response.json().get("detail"), icon=":material/error:")
+
+
+def ensure_valid_access_token():
+    if time.time() >= st.session_state.access_token_expires_at - 30:
+        refresh_tokens()
 
 
 def delete_user(username, password):
@@ -82,6 +103,7 @@ def delete_user(username, password):
 
 
 def create_user_favorite_item(favorite_item: UserFavoriteItemRequest, token: str):
+    ensure_valid_access_token()
     url = f"{BASE_URL}/user_favorite_item/"
     headers = {"Authorization": f"Bearer {token}"}
     payload = favorite_item.dict()
@@ -93,6 +115,7 @@ def create_user_favorite_item(favorite_item: UserFavoriteItemRequest, token: str
 
 
 def get_favorite_items_list(token: str):
+    ensure_valid_access_token()
     url = f"{BASE_URL}/user_favorite_item/"
     headers = {"Authorization": f"Bearer {token}"}
     response = requests.get(url, headers=headers)
@@ -103,6 +126,7 @@ def get_favorite_items_list(token: str):
 
 
 def delete_user_favorite_item(user_favorite_item_request: UserFavoriteItemRequest, token: str):
+    ensure_valid_access_token()
     url = f"{BASE_URL}/user_favorite_item/"
     payload = user_favorite_item_request.dict()
     headers = {"Authorization": f"Bearer {token}"}
@@ -114,6 +138,7 @@ def delete_user_favorite_item(user_favorite_item_request: UserFavoriteItemReques
 
 
 def update_item_quantity(item_order_request: dict, token: str):
+    ensure_valid_access_token()
     url = f"{BASE_URL}/order/"
     headers = {"Authorization": f"Bearer {token}"}
     response = requests.put(url, headers=headers, json=item_order_request)
@@ -124,6 +149,7 @@ def update_item_quantity(item_order_request: dict, token: str):
 
 
 def temp_order_create_and_add_items(order_request: OrderRequest, token: str):
+    ensure_valid_access_token()
     url = f"{BASE_URL}/order/"
     payload = order_request.dict()
     headers = {"Authorization": f"Bearer {token}"}
@@ -135,6 +161,7 @@ def temp_order_create_and_add_items(order_request: OrderRequest, token: str):
 
 
 def close_order(token: str, shipping_address):
+    ensure_valid_access_token()
     url = f"{BASE_URL}/order/close_order"
     headers = {"Authorization": f"Bearer {token}"}
     response = requests.post(url, headers=headers, params={"shipping_address": shipping_address})
@@ -147,6 +174,7 @@ def close_order(token: str, shipping_address):
 
 
 def get_temp_order(token: str):
+    ensure_valid_access_token()
     url = f"{BASE_URL}/order/"
     headers = {"Authorization": f"Bearer {token}"}
     response = requests.get(url, headers=headers)
@@ -157,6 +185,7 @@ def get_temp_order(token: str):
 
 
 def get_all_closed_order(token: str):
+    ensure_valid_access_token()
     url = f"{BASE_URL}/order/get_all_closed_order"
     headers = {"Authorization": f"Bearer {token}"}
     response = requests.get(url, headers=headers)
@@ -167,6 +196,7 @@ def get_all_closed_order(token: str):
 
 
 def del_item_from_temp_order(item_id, token: str):
+    ensure_valid_access_token()
     url = f"{BASE_URL}/order/del_item_from_temp_order"
     headers = {"Authorization": f"Bearer {token}"}
     response = requests.delete(url, headers=headers, params={"item_id": item_id})
@@ -177,6 +207,7 @@ def del_item_from_temp_order(item_id, token: str):
 
 
 def get_user_role(token: str):
+    ensure_valid_access_token()
     url = f"{BASE_URL}/user/role"
     headers = {"Authorization": f"Bearer {token}"}
     response = requests.get(url, headers=headers)
@@ -187,6 +218,7 @@ def get_user_role(token: str):
 
 
 def get_all_users(token: str):
+    ensure_valid_access_token()
     url = f"{BASE_URL}/user/"
     headers = {"Authorization": f"Bearer {token}"}
     response = requests.get(url, headers=headers)
@@ -197,6 +229,7 @@ def get_all_users(token: str):
 
 
 def predict_user_expenses_for_tech_items(user_id: int, token: str):
+    ensure_valid_access_token()
     url = f"{BASE_URL}/predict/predict_user_expenses_for_tech_items"
     headers = {"Authorization": f"Bearer {token}"}
     params = {"user_id": user_id}
@@ -209,6 +242,6 @@ def predict_user_expenses_for_tech_items(user_id: int, token: str):
 
 def chat(session_id, chat_input):
     url = f"{BASE_URL}/chat/ask"
-    response = requests.post(url, json={"session_id": session_id ,"message": chat_input})
+    response = requests.post(url, json={"session_id": session_id, "message": chat_input})
     full_message = response.json()["answer"]
     return full_message
